fix: IndexNotReadyException when processing LSP events
Fixes #711
