+++ src/main/java/com/redhat/devtools/intellij/qute/lsp/QuteLanguageClient.java
@@ -12,11 +12,6 @@ package com.redhat.devtools.intellij.qute.lsp;
 
 import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.module.Module;
-import com.intellij.openapi.progress.ProgressIndicator;
-import com.intellij.openapi.progress.ProgressManager;
-import com.intellij.openapi.progress.Task;
-import com.intellij.openapi.project.DumbService;
-import com.intellij.openapi.project.IndexNotReadyException;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.libraries.Library;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -24,7 +19,7 @@ import com.intellij.util.messages.MessageBusConnection;
 import com.redhat.devtools.intellij.lsp4mp4ij.psi.internal.core.ls.PsiUtilsLSImpl;
 import com.redhat.devtools.intellij.quarkus.QuarkusModuleUtil;
 import com.redhat.devtools.intellij.quarkus.QuarkusProjectService;
-import com.redhat.devtools.intellij.quarkus.lsp4ij.LanguageClientImpl;
+import com.redhat.devtools.intellij.quarkus.lsp4ij.IndexAwareLanguageClient;
 import com.redhat.devtools.intellij.qute.psi.QuteSupportForJava;
 import com.redhat.devtools.intellij.qute.psi.QuteSupportForTemplate;
 import com.redhat.qute.commons.JavaTypeInfo;
@@ -51,17 +46,15 @@ import org.eclipse.lsp4j.CodeLens;
 import org.eclipse.lsp4j.DocumentLink;
 import org.eclipse.lsp4j.Location;
 import org.eclipse.lsp4j.PublishDiagnosticsParams;
-import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class QuteLanguageClient extends LanguageClientImpl implements QuteLanguageClientAPI, QuarkusProjectService.Listener {
+public class QuteLanguageClient extends IndexAwareLanguageClient implements QuteLanguageClientAPI, QuarkusProjectService.Listener {
   private static final Logger LOGGER = LoggerFactory.getLogger(QuteLanguageClient.class);
 
   private final MessageBusConnection connection;
@@ -94,46 +87,6 @@ public class QuteLanguageClient extends LanguageClientImpl implements QuteLangua
             collect(Collectors.toSet());
     if (!uris.isEmpty()) {
       sendPropertiesChangeEvent(uris);
-    }
-  }
-
-  <R> CompletableFuture<R> runAsBackground(String title, Function<ProgressIndicator, R> function) {
-    CompletableFuture<R> future = new CompletableFuture<>();
-    CompletableFuture.runAsync(() -> {
-      Runnable task = () -> ProgressManager.getInstance().run(new Task.Backgroundable(getProject(), title) {
-        @Override
-        public void run(@NotNull ProgressIndicator indicator) {
-          runTask(indicator, future, function);
-        }
-
-        @Override
-        public boolean isHeadless() {
-          return true;
-        }
-      });
-      if (DumbService.getInstance(getProject()).isDumb()) {
-        DumbService.getInstance(getProject()).runWhenSmart(task);
-      } else {
-        task.run();
-      }
-    });
-    return future;
-  }
-
-  private <R> void runTask(@NotNull ProgressIndicator indicator, CompletableFuture<R> future, Function<ProgressIndicator, R> function) {
-    boolean done = false;
-    for(int i=0; !done && i < 10;++i) {
-      try {
-        future.complete(function.apply(indicator));
-        done = true;
-      } catch (IndexNotReadyException e) {
-      } catch (Throwable t) {
-        future.completeExceptionally(t);
-        done = true;
-      }
-    }
-    if (!done) {
-      DumbService.getInstance(getProject()).runWhenSmart(() -> runTask(indicator, future, function));
     }
   }
